# openapi_schema.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CosyVoice API",
    "description": "Cross-lingual Voice Cloning API with CosyVoice2 (v2) and CosyVoice3 (v3)",
    "version": "3.0.0",
    "contact": {
      "name": "CosyVoice API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8012",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v3/streaming/cross-lingual": {
      "post": {
        "tags": ["Streaming (v3)"],
        "summary": "HTTP streaming synthesis",
        "operationId": "stream_synthesis_v3",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/StreamingSynthesisRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Streaming audio",
            "content": {
              "audio/wav": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/streaming/cross-lingual/sse": {
      "get": {
        "tags": ["Streaming (v3)"],
        "summary": "SSE streaming synthesis",
        "operationId": "stream_sse_v3",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "voice_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/AudioFormat"
            }
          },
          {
            "name": "speed",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 1.0
            }
          },
          {
            "name": "quality",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/StreamingQuality"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioFormat": {
        "type": "string",
        "enum": ["wav", "mp3", "flac"],
        "default": "wav"
      },
      "StreamingQuality": {
        "type": "string",
        "enum": ["low", "medium", "high"],
        "default": "medium"
      },
...
      "StreamingSynthesisRequest": {
        "type": "object",
        "required": ["text", "voice_id"],
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000
          },
          "voice_id": {
            "type": "string"
          },
          "format": {
            "$ref": "#/components/schemas/AudioFormat"
          },
          "speed": {
            "type": "number",
            "default": 1.0,
            "minimum": 0.5,
            "maximum": 2.0
          },
          "quality": {
            "$ref": "#/components/schemas/StreamingQuality"
          },
          "chunk_size": {
            "type": "integer",
            "default": 1024,
            "minimum": 256,
            "maximum": 8192
          }
        }
      },
...
    }
  }
}